Highlight tabs + disable UIs based on route.
The tabs in the top bar are now highlighting purely based on the route.
Handles browser refreshes correctly.
At certains routes no tab should be highlighted. For example,
when on the non-course chat pages.

The tab bar and 'peers' pane is also disabled when on
the overview dashboard, since they are not applicable.